+++ packages/client/src/client.ts
@@ -36,7 +36,6 @@ import {
 import { createConstantEntry } from "./constants"
 import { ChainDefinition } from "./descriptors"
 import { createEventEntry } from "./event"
-import { getNewBlocks$, getPinnedBlocks$ } from "./pinnedBlocks"
 import { createRuntimeCallEntry } from "./runtime-call"
 import { createStorageEntry } from "./storage"
 import { createTxEntry, submit, submit$ } from "./tx"
@@ -281,9 +280,7 @@ export function createClient(provider: JsonRpcProvider): PolkadotClient {
   return {
     getChainSpecData,
 
-    blocks$: getNewBlocks$(chainHead.pinnedBlocks$),
-
-    pinnedBlocks$: getPinnedBlocks$(chainHead.pinnedBlocks$),
+    blocks$: chainHead.newBlocks$,
 
     finalizedBlock$: chainHead.finalized$,
     getFinalizedBlock: () => firstValueFrom(chainHead.finalized$),